refactor(runner): isolate lab validation dependency sync
